==> notebooks/vacancy_model/3.0 Combine data.ipynb <==
# 3.0 Combine data

import pandas as pd
from sklearn.ensemble import RandomForestClassifier
import matplotlib.pyplot as plt
import numpy as np
import seaborn as sns
import math
from sklearn.preprocessing import Imputer
from sklearn.base import TransformerMixin
from sklearn import cross_validation
from sklearn.grid_search import GridSearchCV
from sklearn.linear_model import LogisticRegression
from sklearn import preprocessing
from sklearn.cross_validation import cross_val_score
import os

def sound():
    os.system('afplay /System/Library/Sounds/Sosumi.aiff')

%matplotlib inline
sns.set_context('notebook')

path = '/'.join(os.getcwd().split('/')[:-2])
print(path)

df0 = pd.read_csv(path+'/data/model_data/tci_2_0_model.csv', dtype={'Parcel ID':str}, index_col='parcel')
df1 = pd.read_csv(path+'/data/model_data/tci_2_1_model.csv', dtype={'Parcel ID':str}, index_col='parcel')
df2 = pd.read_csv(path+'/data/model_data/tci_2_2_model.csv', dtype={'Parcel ID':str}, index_col='parcel')
df3 = pd.read_csv(path+'/data/model_data/tci_2_3_model.csv', dtype={'Parcel ID':str}, index_col='parcel')

cols = ['Date','vacant','Survey Category','SPA_NAME','Survey Grade']

df = pd.merge(df0, df1[[col for col in df1.columns if col not in cols]], how='inner', left_index=True, right_index=True)
df = pd.merge(df, df2[[col for col in df2.columns if col not in cols]], how='inner', left_index=True, right_index=True)
df = pd.merge(df, df3[[col for col in df3.columns if col not in cols]], how='inner', left_index=True, right_index=True)

df.info()

df['price_per_sqft'] = df['tmktval']/df['totusabl']

# for var in cols2:
#     median_val = dict(df.loc[df[var].notnull(),[var,'SPA_NAME']].groupby('SPA_NAME').agg(np.median)[var])
#     df.loc[df[var].isnull(), var] = df.loc[df[var].isnull()]\
#         .apply(lambda x: median_val[x.SPA_NAME] if x.SPA_NAME in median_val else np.median(df.loc[df[var].notnull()]), axis=1)

# # df['missing_values'] = df.apply(lambda x: len(x[x.isnull()]),axis=1)
# to_remove = ['Parcel ID','parcel','Date','Survey Category','Survey Grade',\
#              'condition13', 'condition14', 'al_date','SPA_NAME','NAME10','style_filtered','rextwall','LENDER_PROCESS_TYPE']

# features = list(df.columns.values)
# for i in to_remove:
#     if i in features:
#         features.remove(i)

# from sklearn import tree

# clf = tree.DecisionTreeClassifier(max_depth=2,min_samples_leaf=5)
# # iris = load_iris()
# clf = clf.fit(df3[features], df3.vacant)
# tree.export_graphviz(clf, out_file=path+'/tree.dot',feature_names=features, impurity=False, label=None,filled=True,class_names=['Not vacant','Vacant']) 

df.to_csv(path+'/data/model_data/tci_final_model.csv')

df.shape
